Extract ContactDetails, Decisions and Course data to their own modules
