+++ app/presenters/vendor_api/application_presenter.rb
@@ -7,6 +7,9 @@ module VendorAPI
     include VendorAPI::WorkExperience
     include VendorAPI::HesaIttData
     include VendorAPI::CandidateData
+    include VendorAPI::CourseData
+    include VendorAPI::DecisionsData
+    include VendorAPI::ContactDetailsData
 
     API_APPLICATION_STATES = { offer_withdrawn: 'rejected',
                                interviewing: 'awaiting_provider_decision' }.freeze
@@ -81,80 +84,11 @@ module VendorAPI
         "What is your subject knowledge?: #{application_form.subject_knowledge}"
     end
 
-    def contact_details
-      if application_form.international_address?
-        address_line1 = application_form.address_line1 || application_form.international_address
-
-        {
-          phone_number: application_form.phone_number,
-          address_line1: address_line1,
-          address_line2: application_form.address_line2,
-          address_line3: application_form.address_line3,
-          address_line4: application_form.address_line4,
-          country: application_form.country,
-          email: application_form.candidate.email_address,
-        }
-      else
-        {
-          phone_number: application_form.phone_number,
-          address_line1: application_form.address_line1,
-          address_line2: application_form.address_line2,
-          address_line3: application_form.address_line3,
-          address_line4: application_form.address_line4,
-          postcode: application_form.postcode,
-          country: application_form.country,
-          email: application_form.candidate.email_address,
-        }
-      end
-    end
-
-      {
-        conditions: application_choice.offer.conditions_text,
-        offer_made_at: application_choice.offered_at,
-        offer_accepted_at: application_choice.accepted_at,
-        offer_declined_at: application_choice.declined_at,
-      }.merge(current_course)
-    end
-
     def references
       references = application_form.application_references
       references.select { |reference| reference.selected && reference.feedback_provided? }.map do |reference|
         reference_to_hash(reference)
       end
-    end
-
-    def rejection
-      @rejection ||= if application_choice.rejection_reason? || application_choice.structured_rejection_reasons.present?
-                       {
-                         reason: VendorAPI::RejectionReasonPresenter.new(application_choice).present,
-                         date: application_choice.rejected_at.iso8601,
-                       }
-                     elsif application_choice.offer_withdrawal_reason?
-                       {
-                         reason: application_choice.offer_withdrawal_reason,
-                         date: application_choice.offer_withdrawn_at.iso8601,
-                       }
-                     elsif application_choice.rejected_by_default?
-                       {
-                         reason: 'Not entered',
-                         date: application_choice.rejected_at.iso8601,
-                       }
-                     end
-      return if @rejection.blank?
-
-      {
-        reason: truncate_if_over_advertised_limit('Rejection.properties.reason', @rejection[:reason]),
-        date: @rejection[:date],
-      }
-    end
-
-    def withdrawal
-      return unless application_choice.withdrawn?
-
-      {
-        reason: nil, # Candidates are not able to provide a withdrawal reason yet
-        date: application_choice.withdrawn_at.iso8601,
-      }
     end
 
     def safeguarding_issues_details_url
