+++ spec/presenters/vendor_api/application_presenter_spec.rb
@@ -120,133 +120,6 @@ RSpec.describe VendorAPI::ApplicationPresenter do
     end
   end
 
-  describe '#withdrawal' do
-    let(:withdrawn_at) { Time.zone.local(2019, 1, 1, 0, 0, 0) }
-    let!(:application_choice) { create(:application_choice, :withdrawn, application_form: application_form, withdrawn_at: withdrawn_at) }
-
-    it 'returns a withdrawal object' do
-      expect(attributes[:withdrawal]).to eq({ reason: nil, date: withdrawn_at.iso8601 })
-    end
-  end
-
-  describe '#rejection' do
-    let(:rejected_at) { Time.zone.local(2019, 1, 1, 0, 0, 0) }
-    let!(:application_choice) { create(:application_choice, :rejected, application_form: application_form, rejected_at: rejected_at, rejection_reason: 'Course full') }
-
-    it 'returns a rejection object' do
-      expect(attributes[:rejection]).to eq({ reason: 'Course full', date: rejected_at.iso8601 })
-    end
-
-    it 'returns a rejection object with a truncated reason when the character limit is exceeded' do
-      application_choice.rejection_reason = 'Course full' * 65000
-      allow(Sentry).to receive(:capture_message)
-
-      application_json
-
-      expect(Sentry).to have_received(:capture_message).with("Rejection.properties.reason truncated for application with id #{application_choice.id} as length exceeded 65535 chars")
-
-      expect(attributes[:rejection][:reason].length).to be(65535)
-      expect(attributes[:rejection][:reason]).to end_with(described_class::OMISSION_TEXT)
-      expect(attributes[:rejection][:date]).to eq(rejected_at.iso8601)
-    end
-
-    context 'when there is a withdrawn offer' do
-      let(:withdrawn_at) { Time.zone.local(2019, 1, 1, 0, 0, 0) }
-      let(:application_choice) { create(:application_choice, :rejected, application_form: application_form, offer_withdrawn_at: withdrawn_at, offer_withdrawal_reason: 'Course full') }
-
-      it 'returns a rejection object' do
-        expect(attributes[:rejection]).to eq({ reason: 'Course full', date: withdrawn_at.iso8601 })
-      end
-    end
-
-    context 'when there is no feedback' do
-      let(:application_choice) { create(:application_choice, :with_rejection_by_default, application_form: application_form, rejected_at: rejected_at) }
-
-      it 'returns a rejection object with a custom rejection reason' do
-        expect(attributes[:rejection]).to eq({ reason: 'Not entered', date: rejected_at.iso8601 })
-      end
-    end
-  end
-
-  describe '#contact_details' do
-    let(:application_choice) { create(:application_choice, :awaiting_provider_decision, application_form: application_form) }
-    let(:application_form) { create(:application_form, :minimum_info, application_form_attributes) }
-    let(:application_form_attributes) do
-      {
-        phone_number: '07700 900 982',
-        address_line1: '42',
-        address_line2: 'Much Wow Street',
-        address_line3: 'London',
-        address_line4: 'England',
-        country: 'GB',
-        postcode: 'SW1P 3BT',
-      }
-    end
-
-    context 'for UK addresses' do
-      it 'returns contact details in correct format' do
-        expected_contact_details = application_form_attributes.merge(email: application_form.candidate.email_address)
-
-        expect(attributes[:contact_details]).to eq(expected_contact_details)
-      end
-    end
-
-    context 'for international addresses' do
-      let(:application_form_attributes) do
-        {
-          phone_number: '07700 900 982',
-          address_type: 'international',
-          address_line1: '456 Marine Drive',
-          address_line2: 'Mumbai',
-          address_line3: nil,
-          address_line4: nil,
-          international_address: '456 Marine Drive, Mumbai',
-          country: 'IN',
-        }
-      end
-
-      it 'returns contact details in correct format' do
-        expected_contact_details = {
-          phone_number: '07700 900 982',
-          address_line1: '456 Marine Drive',
-          address_line2: 'Mumbai',
-          address_line3: nil,
-          address_line4: nil,
-          country: 'IN',
-        }.merge(email: application_form.candidate.email_address)
-
-        expect(attributes[:contact_details]).to eq(expected_contact_details)
-      end
-    end
-
-    context 'if no address lines are populated' do
-      let(:application_form_attributes) do
-        {
-          phone_number: '07700 900 982',
-          address_type: 'international',
-          international_address: '456 Marine Drive, Mumbai',
-          address_line1: nil,
-          address_line2: nil,
-          address_line3: nil,
-          address_line4: nil,
-          country: 'IN',
-        }
-      end
-
-      it 'presents the international_address field' do
-        expect(attributes[:contact_details]).to eq({
-          phone_number: '07700 900 982',
-          address_line1: '456 Marine Drive, Mumbai',
-          address_line2: nil,
-          address_line3: nil,
-          address_line4: nil,
-          country: 'IN',
-          email: application_form.candidate.email_address,
-        })
-      end
-    end
-  end
-
   describe '#references' do
     let(:application_choice) { create(:application_choice, :with_completed_application_form, :with_offer) }
 
@@ -271,44 +144,6 @@ RSpec.describe VendorAPI::ApplicationPresenter do
       it 'are mapped on the reference object' do
         expect(attributes[:references].map { |reference| reference[:safeguarding_concerns] })
           .to match_array [true, false]
-      end
-    end
-  end
-
-  describe '#offer' do
-    context 'when there is no offer' do
-      let(:application_choice) { create(:application_choice, :with_completed_application_form, :awaiting_provider_decision) }
-
-      it 'returns nil' do
-        expect(attributes[:offer]).to be_nil
-      end
-    end
-
-    context 'when there is an offer' do
-      let(:application_choice) { create(:application_choice, :with_completed_application_form, :with_offer) }
-
-      it 'includes an offer_made_at date for offers' do
-        expect(attributes[:offer][:offer_made_at]).to be_present
-      end
-
-      it 'includes the offered course' do
-        expect(attributes[:offer][:course][:course_code]).to eq(application_choice.current_course_option.course.code)
-      end
-    end
-
-    context 'when there is an accepted offer' do
-      let(:application_choice) { create(:application_choice, :with_completed_application_form, :with_accepted_offer) }
-
-      it 'includes an accepted_at date for accepted offers' do
-        expect(attributes[:offer][:offer_accepted_at]).to be_present
-      end
-    end
-
-    context 'when there is a declined offer' do
-      let(:application_choice) { create(:application_choice, :with_completed_application_form, :with_declined_offer) }
-
-      it 'includes a declined_at date for declined offers' do
-        expect(attributes[:offer][:offer_declined_at]).to be_present
       end
     end
   end
